Better error handling
Adds use lines for the exceptions which fixes an error that breaks
the dashboard, and change existing exceptions to an ApplicationException
so that the errors don't get logged. Otherwise a permission error would
be logged every time the dashboard is loaded.

Also adds extra checking for the log so we can throw a more specific error
message if the whole thing fails to load entirely.